backend/requirements.txt
fastapi==0.79.0
uvicorn[standard]==0.18.2


# PyJWT==1.7.1
python-jose==3.3.0
python-decouple==3.6
python-multipart==0.0.5
passlib[bcrypt]==1.7.4 
pydantic[email]==1.9.1